Create README.md
It was actually developed overnight for Spirathon. 
It was designed so that people can report any diseases that might be appearing around their locality.
It uses the location data that is sent to the server while making reports to notify people around that region.
Currently it is not functional as it requires a server which is no longer running after the competition.